New & Updated Icons
28 New Icons added
Updated Wikipedia, PayPal, and Itch.io Icons
Removed 2 unnecessary dark icons
Removed Unlockbase icon as it does not meet the criteria to have an icon.

On the 2fa.directory we now fully cover the services on the following categories that meet our criteria for an icon:
- Backup & Sync
- Banking
- Betting
- Cloud Computing
- Crowdfunding
- Food
- Government
- Health
- Task Management
- Transport

We already have large coverage over the other categories and I will continue to work  my way through them.